requirements.txt
tiktoken
dotenv-python